+++ local.properties
@@ -4,5 +4,5 @@
 # Location of the SDK. This is only used by Gradle.
 # For customization when using a Version Control System, please read the
 # header note.
-#Thu Apr 10 20:29:51 IDT 2025
-sdk.dir=C\:\\Users\\PC\\AppData\\Local\\Android\\Sdk
+#Thu Jul 03 18:09:51 IDT 2025
+sdk.dir=C\:\\Users\\yalia\\AppData\\Local\\Android\\Sdk
